Modernize (part 2): macOS dispatch semaphores + warning cleanup
- sr_compat.h: use GCD dispatch semaphores on macOS instead of the
  deprecated (and unsupported) unnamed POSIX sem_init/sem_wait/sem_post.
  Removes the last -Wdeprecated-declarations warning and gives macOS
  actually-working semaphores. Linux/other-unix still use POSIX sems.
- Clear -Wall/-Wextra findings that were real (not the intentional
  callback unused-params):
  - remove dead variables (ripstream write_list/mfullpath/fullpath,
    utf8 charset, filelib done flag);
  - keep side-effecting calls, drop unused result vars (findsep/findsep2
    mad_decoder_run, filelib get_next_sequence_number string_from_gstring);
  - fix sign-compare casts (charset, cbuf3, mchar, ripstream); also fix a
    latent bug in charset.c where (1<<31) overflowed to a negative int.
- CMakeLists: drop duplicate -lm (silences a linker warning).

No behavior change on Linux; macOS semaphores now functional. Verified
mp3, HLS, and the track-split pipeline still work. Still 3.0.0-qbus.

+++ lib/cbuf3.c
@@ -92,7 +92,7 @@ cbuf3_allocate_minimum (struct cbuf3 *cbuf3,
 	return SR_SUCCESS;
     }
 
-    for (i = 0; i < num_chunks - cbuf3->num_chunks; i++) {
+    for (i = 0; (unsigned long) i < num_chunks - cbuf3->num_chunks; i++) {
 	char* chunk = (char*) malloc (cbuf3->chunk_size);
 	if (!chunk) {
 	    threadlib_signal_sem (&cbuf3->sem);
